Split files into single responsibility

+++ gpt3_assistant/computer_voice.py
@@ -5,7 +5,7 @@ import logging
 
 
 class ComputerVoice:
-    def __init__(self, mp3_filename="temp.mp3", lang="en", tld="com"):
+    def __init__(self, mp3_filename, lang=None, tld=None):
         self._mp3_filename = mp3_filename
         self._language = lang
         self._top_level_domain = tld
@@ -20,21 +20,34 @@ class ComputerVoice:
             logging.error(f"Error cleaning up ComputerVoice context: {e}")
 
     def speak(self, text_to_speak: str):
+        """
+        Speak the referenced text on the machine speakers.
+        :param text_to_speak: the text to speak.
+        :return: None
+        """
         logging.debug(f"ComputerVoice.speak - '{text_to_speak}'")
-        tts = self.get_gtts(text_to_speak)
+        tts = self._get_gtts(text_to_speak)
         tts.save(self._mp3_filename)
         full_mp3_path = os.path.join(os.getcwd(), self._mp3_filename)
         subprocess.call(["afplay", full_mp3_path])
 
     def cleanup_temp_files(self):
+        """
+        Remove all temporary files and cleanup before shutting down.
+        :return: None
+        """
         logging.debug(f"ComputerVoice.cleanup_temp_files - {self._mp3_filename}")
-        os.remove(self._mp3_filename)
 
-    def get_gtts(self, text_to_speak: str):
+        # check if temporary file exists before trying to delete
+        if os.path.exists(self._mp3_filename):
+            os.remove(self._mp3_filename)
+
+    def _get_gtts(self, text_to_speak: str):
         logging.debug(
             f"ComputerVoice.get_gtts - Language: {self._language}, TLD: {self._top_level_domain}"
         )
         exception_thrown = True
+
         try:
             gtts_instance = None
 
@@ -46,7 +59,7 @@ class ComputerVoice:
                     text_to_speak, lang=self._language, tld=self._top_level_domain
                 )
             else:
-                logging.debug("Using default language: English (United States)")
+                logging.debug("Using default language")
                 gtts_instance = gtts.gTTS(text_to_speak)
 
             exception_thrown = False
